requirements.txt
altair==4.1.0
altair-data-server==0.4.1
altair-saver==0.5.0
altair-viewer==0.3.0
atomicwrites==1.4.0
beautifulsoup4==4.9.1
bs4==0.0.1
cmasher==1.5.5
cycler==0.10.0
emoji==0.6.0
Flask
Flask-Caching==1.9.0
Flask-Cors==3.0.9
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
gensim==3.8.3
google-api-core==1.22.4
google-api-python-client==1.12.3
google-auth==1.22.1
google-auth-httplib2==0.0.4
google-play-scraper==0.1.1
googleapis-common-protos==1.52.0
googletrans==3.0.0
gunicorn==19.10.0
ipywidgets==7.5.1
langdetect==1.0.8
matplotlib==3.2.2
nltk==3.5
notebook==6.1.4
numpy==1.19.3
oauth2client==4.1.3
plac==1.1.3
py==1.9.0
python-dotenv
regex==2020.6.8
reportlab==3.5.53
seaborn==0.10.1
semver==2.10.2
statsmodels==0.11.1
testpath==0.4.4
textblob==0.15.3
tqdm==4.47.0
Werkzeug==1.0.1
wordcloud==1.8.0
WTForms==2.3.3
